requirements.txt
openai==0.27.0
requests~=2.31.0
tenacity~=8.2.2
termcolor~=2.3.0
boto3~=1.26.165
pydub~=0.25.1
SpeechRecognition
python_dotenv
pyaudio~=0.2.13
python-dotenv~=1.0.0
EdgeGPT~=0.12.0
eventregistry~=9.1
spacy~=3.6.0
asyncio~=3.4.3
yt-dlp~=2023.7.6
openai~=0.27.0
openai-whisper~=20230314
-f https://download.pytorch.org/whl/torch_stable.html
torch
torchaudio~=2.0.2
torchvision~=0.15.2
beautifulsoup4~=4.12.2


